Remove long-running unstable test
This exodiff'd on code changes that I would expect to be unrelated while
other tests that use the exact same set of objects passed. So I consider
this test to be a bit unstable. Moreover it takes a minute and a half to
run. Since we run the exact input file, just with a different `xmax` and
`ymax`, in a different test, I feel valid in removing this test